refactor(pneu): altera tipo do campo custoAquisicao
PL-3459

+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/v3/_model/dto/PneuCadastro.java
@@ -45,7 +45,7 @@ public class PneuCadastro {
     String dot;
 
     @NotNull(message = "O custo de aquisição não pode ser nulo.")
-    Long custoAquisicao;
+    Double custoAquisicao;
 
     boolean pneuUsado;
